+++ src/store/user-data.ts
@@ -65,6 +65,7 @@ export const initialUserData = deepFreeze({
   tabs: {
     order: [] as string[],
     hidden: [] as string[],
+    locked: [] as string[],
   },
   commandLists: [] as UserData.CommandList[],
 });